Moved logic to fetch plans from WPMainActivity to MySiteFragment, it is only needed in MySiteFragment so it makes more sense to be contained within that fragment
